\name{eigentest-package}
\alias{eigentest-package}
\alias{eigentest}
\docType{package}
\title{
\packageTitle{eigentest}
}
\description{
\packageDescription{eigentest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{eigentest}
\packageIndices{eigentest}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{eigentest}

Maintainer: \packageMaintainer{eigentest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
